node-neigh: Avoid arpinging the same next hop
When nodes are connected via L3 GW, the periodic ARP refresher pings the
same IP (GW). To avoid that, introduce neighLastPingByNextHop which
stores a timestamp of the last successful arping for a given next hop.
If delta between the last and the next arping is less than
option.Config.ARPPingRefreshPeriod, then the arping is skipped until the
next time the refresher is scheduled.

Signed-off-by: Martynas Pumputis <m@lambda.lt>